Crate embedded_nal

embedded-nal - A Network Abstraction Layer for Embedded Systems

Re-exports

pub use nb;

Structs

SharableStack

Sharable wrapper for a network stack implementation.

SharedStack

Single-thread shared reference to an internal network stack implementation.

Enums

AddrType

This is the host address type to be returned by gethostbyname.

TcpErrorKind

Represents specific errors encountered during TCP operations.

Traits

Dns

This trait is an extension trait for TcpStack and UdpStack for dns resolutions. It does not handle every DNS record type, but is meant as an embedded alternative to ToSocketAddrs, and is as such meant to resolve an ip address from a hostname, or a hostname from an ip address. This means that it only deals in host address records A (IPv4) and AAAA (IPv6).

TcpClientStack

This trait is implemented by TCP/IP stacks. You could, for example, have an implementation which knows how to send AT commands to an ESP8266 WiFi module. You could have another implementation which knows how to driver the Rust Standard Library’s std::net module. Given this trait, you can write a portable HTTP client which can work with either implementation.

TcpError

Methods to resolve errors into identifiable, actionable codes on the client side.

TcpFullStack

This trait is implemented by TCP/IP stacks that expose TCP server functionality. TCP servers may listen for connection requests to establish multiple unique TCP connections with various clients.

UdpClientStack

This trait is implemented by UDP/IP stacks. You could, for example, have an implementation which knows how to send AT commands to an ESP8266 WiFi module. You could have another implementation which knows how to driver the Rust Standard Library’s std::net module. Given this trait, you can how write a portable CoAP client which can work with either implementation.

UdpFullStack

This trait is implemented by UDP/IP stacks. It provides the ability to listen for packets on a specified port and send replies.