Struct embedded_nal::Ipv4Addr

pub struct Ipv4Addr { /* private fields */ }

An IPv4 address.

IPv4 addresses are defined as 32-bit integers in IETF RFC 791. They are usually represented as four octets.

See IpAddr for a type encompassing both IPv4 and IPv6 addresses.

Textual representation

Ipv4Addr provides a FromStr implementation. The four octets are in decimal notation, divided by . (this is called “dot-decimal notation”). Notably, octal numbers and hexadecimal numbers are not allowed per IETF RFC 6943.

Examples

use no_std_net::Ipv4Addr;

let localhost = Ipv4Addr::new(127, 0, 0, 1);
assert_eq!("127.0.0.1".parse(), Ok(localhost));
assert_eq!(localhost.is_loopback(), true);

Implementations

impl Ipv4Addr

pub const fn new(a: u8, b: u8, c: u8, d: u8) -> Ipv4Addr

Creates a new IPv4 address from four eight-bit octets.

The result will represent the IP address a.b.c.d.

Examples

use no_std_net::Ipv4Addr;

let addr = Ipv4Addr::new(127, 0, 0, 1);

pub const LOCALHOST: Ipv4Addr = _

An IPv4 address with the address pointing to localhost: 127.0.0.1

Examples

use no_std_net::Ipv4Addr;

let addr = Ipv4Addr::LOCALHOST;
assert_eq!(addr, Ipv4Addr::new(127, 0, 0, 1));

pub const UNSPECIFIED: Ipv4Addr = _

An IPv4 address representing an unspecified address: 0.0.0.0

This corresponds to the constant INADDR_ANY in other languages.

Examples

use no_std_net::Ipv4Addr;

let addr = Ipv4Addr::UNSPECIFIED;
assert_eq!(addr, Ipv4Addr::new(0, 0, 0, 0));

pub const BROADCAST: Ipv4Addr = _

An IPv4 address representing the broadcast address: 255.255.255.255

Examples

use no_std_net::Ipv4Addr;

let addr = Ipv4Addr::BROADCAST;
assert_eq!(addr, Ipv4Addr::new(255, 255, 255, 255));

pub const fn octets(&self) -> [u8; 4]

Returns the four eight-bit integers that make up this address.

Examples

use no_std_net::Ipv4Addr;

let addr = Ipv4Addr::new(127, 0, 0, 1);
assert_eq!(addr.octets(), [127, 0, 0, 1]);

pub const fn is_unspecified(&self) -> bool

Returns true for the special ‘unspecified’ address (0.0.0.0).

This property is defined in UNIX Network Programming, Second Edition, W. Richard Stevens, p. 891; see also ip7.

Examples

use no_std_net::Ipv4Addr;

assert_eq!(Ipv4Addr::new(0, 0, 0, 0).is_unspecified(), true);
assert_eq!(Ipv4Addr::new(45, 22, 13, 197).is_unspecified(), false);

pub const fn is_loopback(&self) -> bool

Returns true if this is a loopback address (127.0.0.0/8).

This property is defined by IETF RFC 1122.

Examples

use no_std_net::Ipv4Addr;

assert_eq!(Ipv4Addr::new(127, 0, 0, 1).is_loopback(), true);
assert_eq!(Ipv4Addr::new(45, 22, 13, 197).is_loopback(), false);

pub const fn is_private(&self) -> bool

Returns true if this is a private address.

The private address ranges are defined in IETF RFC 1918 and include:

10.0.0.0/8

172.16.0.0/12

192.168.0.0/16

Examples

use no_std_net::Ipv4Addr;

assert_eq!(Ipv4Addr::new(10, 0, 0, 1).is_private(), true);
assert_eq!(Ipv4Addr::new(10, 10, 10, 10).is_private(), true);
assert_eq!(Ipv4Addr::new(172, 16, 10, 10).is_private(), true);
assert_eq!(Ipv4Addr::new(172, 29, 45, 14).is_private(), true);
assert_eq!(Ipv4Addr::new(172, 32, 0, 2).is_private(), false);
assert_eq!(Ipv4Addr::new(192, 168, 0, 2).is_private(), true);
assert_eq!(Ipv4Addr::new(192, 169, 0, 2).is_private(), false);

pub const fn is_link_local(&self) -> bool

Returns true if the address is link-local (169.254.0.0/16).

This property is defined by IETF RFC 3927.

Examples

use no_std_net::Ipv4Addr;

assert_eq!(Ipv4Addr::new(169, 254, 0, 0).is_link_local(), true);
assert_eq!(Ipv4Addr::new(169, 254, 10, 65).is_link_local(), true);
assert_eq!(Ipv4Addr::new(16, 89, 10, 65).is_link_local(), false);

pub const fn is_multicast(&self) -> bool

Returns true if this is a multicast address (224.0.0.0/4).

Multicast addresses have a most significant octet between 224 and 239, and is defined by IETF RFC 5771.

Examples

use no_std_net::Ipv4Addr;

assert_eq!(Ipv4Addr::new(224, 254, 0, 0).is_multicast(), true);
assert_eq!(Ipv4Addr::new(236, 168, 10, 65).is_multicast(), true);
assert_eq!(Ipv4Addr::new(172, 16, 10, 65).is_multicast(), false);

pub const fn is_broadcast(&self) -> bool

Returns true if this is a broadcast address (255.255.255.255).

A broadcast address has all octets set to 255 as defined in IETF RFC 919.

Examples

use no_std_net::Ipv4Addr;

assert_eq!(Ipv4Addr::new(255, 255, 255, 255).is_broadcast(), true);
assert_eq!(Ipv4Addr::new(236, 168, 10, 65).is_broadcast(), false);

pub const fn is_documentation(&self) -> bool

Returns true if this address is in a range designated for documentation.

This is defined in IETF RFC 5737:

192.0.2.0/24 (TEST-NET-1)

198.51.100.0/24 (TEST-NET-2)

203.0.113.0/24 (TEST-NET-3)

Examples

use no_std_net::Ipv4Addr;

assert_eq!(Ipv4Addr::new(192, 0, 2, 255).is_documentation(), true);
assert_eq!(Ipv4Addr::new(198, 51, 100, 65).is_documentation(), true);
assert_eq!(Ipv4Addr::new(203, 0, 113, 6).is_documentation(), true);
assert_eq!(Ipv4Addr::new(193, 34, 17, 19).is_documentation(), false);

pub const fn to_ipv6_compatible(&self) -> Ipv6Addr

Converts this address to an IPv4-compatible IPv6 address.

a.b.c.d becomes ::a.b.c.d

This isn’t typically the method you want; these addresses don’t typically function on modern systems. Use to_ipv6_mapped instead.

Examples

use no_std_net::{Ipv4Addr, Ipv6Addr};

assert_eq!(
    Ipv4Addr::new(192, 0, 2, 255).to_ipv6_compatible(),
    Ipv6Addr::new(0, 0, 0, 0, 0, 0, 0xc000, 0x2ff)
);

pub const fn to_ipv6_mapped(&self) -> Ipv6Addr

Converts this address to an IPv4-mapped IPv6 address.

a.b.c.d becomes ::ffff:a.b.c.d

Examples

use no_std_net::{Ipv4Addr, Ipv6Addr};

assert_eq!(Ipv4Addr::new(192, 0, 2, 255).to_ipv6_mapped(),
           Ipv6Addr::new(0, 0, 0, 0, 0, 0xffff, 0xc000, 0x2ff));